requirements.txt
streamlit
langchain
langchain-google-genai
beautifulsoup4
faiss-cpu
python-dotenv